Separate homepage columns container into component for reuse
