+++ components/Container.js
@@ -1,13 +1,23 @@
-import { Flex } from '@chakra-ui/core';
+import { Grid } from '@chakra-ui/core';
 
-function Container(props) {
+function Container({ desktopWidth=85, mobileWidth=95, children }) {
+
+  const desktop = `${desktopWidth}%`;
+  const mobile = `${mobileWidth}%`;
+  const responsiveWidth = [mobile, mobile, desktop, desktop];
+
+  const splitColumns = '60% 40%';
+  const oneColumn = '100%';
+  const responsiveColumns = [oneColumn, oneColumn, oneColumn, splitColumns];
+
   return (
-    <Flex 
-      direction="column"
-      alignItems="center"
-      justifyContent="flex-start"
-      {...props}
-    />
+    <Grid 
+      margin="auto"
+      templateColumns={responsiveColumns} 
+      width={responsiveWidth}
+    >
+      {children}
+    </Grid>
   );
 }
