+++ pages/index.js
@@ -2,21 +2,13 @@ import Link from "next/link";
 import { Box, Grid, Heading, List, ListItem } from "@chakra-ui/core";
 import NavigationBar from "../components/NavigationBar";
 import Hero from "../components/Hero";
-import Builds from '../components/Builds';
+import Container from "../components/Container";
+import Builds from "../components/Builds";
 import { fetchBuilds, fetchUsers, fetchTags } from "../clients";
 
 // TODO(Renzo): Create components for displaying builds and users
 
 export default function Home({ builds, users, tags }) {
-
-  const desktopWidth = '85%';
-  const fullWidth = '95%';
-  const responsiveWidth = [fullWidth, fullWidth, desktopWidth, desktopWidth];
-
-  const splitColumns = '60% 40%';
-  const oneColumn = '100%';
-  const responsiveColumns = [oneColumn, oneColumn, oneColumn, splitColumns];
-  
   return (
     <div>
       <NavigationBar />
@@ -24,11 +16,7 @@ export default function Home({ builds, users, tags }) {
         headerText={'Self-teaching done right.'}
         subText={'Start building your way to learn.'}  
       />
-      <Grid 
-        margin="auto"
-        templateColumns={responsiveColumns} 
-        width={responsiveWidth}
-      >
+      <Container>
         <Box>
           <Builds builds={builds} tags={tags} />
         </Box>
@@ -45,7 +33,7 @@ export default function Home({ builds, users, tags }) {
             ))}
           </List>
         </Box>
-      </Grid>
+      </Container>
     </div>
   );
 }
